Build Ship class
I'm not sure if I need to call the isSunk method inside the hit method.
Because that bothers me that you can hit a Ship multiple times and the
sunk property is only changed when the isSunk method is called. It's a
bit like the Schrodinger cat, it is not dead until you lift the box/call
the isSunk method.