Build the generateRandomShipPlacements function
Yeeeees !! I finally did it ! It took some time to gather my courage to
finally get it done. I really thought it was impossible, but I just
needed some time to think and to believe everything is possible. It is
the best part to get stuck on something, it's where you build your
skills and your confidence !! Let's go Ulysse